Add transfer safe tokens SDK and CLI
