+++ .vscode/launch.json
@@ -49,16 +49,16 @@
       "args": [
         // Kovan CARD Bridge args
         // "bridge",
+        // "10000",
         // "0xd6E34821F508e4247Db359CFceE0cb5e8050972a",
-        // "1000",
         // "--network", "kovan",
         // // Hassan's testing mnemonic feel free to use your own
         // "--mnemonic", "pizza monitor radio able holiday boil beyond kingdom throw evil limb dream"
 
-        // Kovan Dai Bridge args
+        // Kovan DAI Bridge args
         "bridge",
         "10",
-        "0x4F96Fe3b7A6Cf9725f59d353F723c1bDb64CA6Aa",
+        "0x4F96Fe3b7A6Cf9725f59d353F723c1bDb64CA6Aa", // Kovan DAI
         "--network", "kovan",
         // Hassan's testing mnemonic feel free to use your own
         "--mnemonic", "pizza monitor radio able holiday boil beyond kingdom throw evil limb dream"
@@ -83,15 +83,32 @@
     {
       "type": "node",
       "request": "launch",
+      "name": "Cardpay: Transfer Tokens from Safe",
+      "program": "${workspaceFolder}/packages/cardpay-cli/cardpay.js",
+      "console": "integratedTerminal",
+      "env": {},
+      "args": [
+        "safe-transfer-tokens",
+        "0xaE9D292753f05b00E996F6b18E4bAe5CC9e5aa48", // Hassan's depot safe feel free to use your own
+        "0xB236ca8DbAB0644ffCD32518eBF4924ba866f7Ee", // CARD.CPXD
+        "0xD7182E380b7dFa33C186358De7E1E5d0950fCAE7", // Relay txn sender
+        "100000",
+        "--network", "sokol",
+        // Hassan's testing mnemonic feel free to use your own
+        "--mnemonic", "pizza monitor radio able holiday boil beyond kingdom throw evil limb dream"
+      ]
+    },
+    {
+      "type": "node",
+      "request": "launch",
       "name": "Cardpay: Create Prepaid Card",
       "program": "${workspaceFolder}/packages/cardpay-cli/cardpay.js",
       "console": "integratedTerminal",
       "env": {},
       "args": [
         "prepaidcard-create",
-        // Hassan's depot safe feel free to use your own
-        "0xaE9D292753f05b00E996F6b18E4bAe5CC9e5aa48",
-        "0xFeDc0c803390bbdA5C4C296776f4b574eC4F30D1",
+        "0xaE9D292753f05b00E996F6b18E4bAe5CC9e5aa48", // Hassan's depot safe feel free to use your own
+        "0xFeDc0c803390bbdA5C4C296776f4b574eC4F30D1", // DAI.CPXD
         "500",
         "--network", "sokol",
         // Hassan's testing mnemonic feel free to use your own
@@ -110,8 +127,7 @@
         // currently creating a merchant is an administrative function--ping
         // hassan to have him create a merchant for you until we get the self-serve fucntionality live
         "0x8B3A993F4f8159A5FcFEe3155a33BD1E9Bd6ABe4", // safe for merchant 0x54b11672e90B38ca52140Ac364EebC5cC07F3d91
-        // Hassan's prepaid card --feel free to use your own
-        "0x6931aDbbefAfFCB0E1F3a4EbA95A5017386c9752",
+        "0x6931aDbbefAfFCB0E1F3a4EbA95A5017386c9752", // Hassan's prepaid card --feel free to use your own
         "100",
         "--network", "sokol",
         // Hassan's testing mnemonic feel free to use your own
@@ -127,7 +143,7 @@
       "env": {},
       "args": [
         "price-for-face-value",
-        "0xFeDc0c803390bbdA5C4C296776f4b574eC4F30D1",
+        "0xFeDc0c803390bbdA5C4C296776f4b574eC4F30D1", // DAI.CPXD
         "10000",
         "--network", "sokol",
         // Hassan's testing mnemonic feel free to use your own
@@ -143,7 +159,7 @@
       "env": {},
       "args": [
         "new-prepaidcard-gas-fee",
-        "0xFeDc0c803390bbdA5C4C296776f4b574eC4F30D1",
+        "0xFeDc0c803390bbdA5C4C296776f4b574eC4F30D1", // DAI.CPXD
         "--network", "sokol",
         // Hassan's testing mnemonic feel free to use your own
         "--mnemonic", "pizza monitor radio able holiday boil beyond kingdom throw evil limb dream"
