+++ packages/cardpay-cli/safe.ts
@@ -1,8 +1,10 @@
 import Web3 from 'web3';
-import { Safes } from '@cardstack/cardpay-sdk';
+import { Safes, Assets, getConstant } from '@cardstack/cardpay-sdk';
 import { getWeb3 } from './utils';
 
-export default async function (network: string, mnemonic: string, address?: string): Promise<void> {
+const { toWei } = Web3.utils;
+
+export async function viewSafes(network: string, mnemonic: string, address?: string): Promise<void> {
   let web3 = await getWeb3(network, mnemonic);
 
   let safesApi = new Safes(web3);
@@ -30,4 +32,25 @@ export default async function (network: string, mnemonic: string, address?: stri
 
     console.log('\n');
   });
+}
+
+export async function transferTokens(
+  network: string,
+  mnemonic: string,
+  safe: string,
+  token: string,
+  recipient: string,
+  amount: number
+): Promise<void> {
+  let web3 = await getWeb3(network, mnemonic);
+  let weiAmount = toWei(String(amount));
+
+  let safes = new Safes(web3);
+  let assets = new Assets(web3);
+  let { symbol } = await assets.getTokenInfo(token);
+
+  console.log(`transferring ${amount} ${symbol} from safe ${safe} to ${recipient}`);
+  let result = await safes.sendTokens(safe, token, recipient, weiAmount);
+  let blockExplorer = await getConstant('blockExplorer', web3);
+  console.log(`Transaction hash: ${blockExplorer}/tx/${result.ethereumTx.txHash}/token-transfers`);
 }
