+++ packages/cardpay-sdk/sdk/safes.ts
@@ -4,10 +4,15 @@ import Web3 from 'web3';
 import PrepaidCardManagerABI from '../contracts/abi/prepaid-card-manager';
 import RevenuePoolABI from '../contracts/abi/revenue-pool';
 import BridgeUtilsABI from '../contracts/abi/bridge-utils';
+import ERC20ABI from '../contracts/abi/erc-20';
 import { AbiItem } from 'web3-utils';
 import { getAddress } from '../contracts/addresses';
 import { getConstant, ZERO_ADDRESS } from './constants';
 import ExchangeRate from './exchange-rate';
+import { ContractOptions } from 'web3-eth-contract';
+import { GnosisExecTx, sign, gasEstimate, executeTransaction } from './utils';
+import BN from 'bn.js';
+const { toBN, fromWei } = Web3.utils;
 
 export type Safe = DepotSafe | PrepaidCardSafe | MerchantSafe | ExternalSafe;
 interface BaseSafe {
@@ -106,5 +111,66 @@ export default class Safes {
         };
       })
     );
+  }
+
+  async sendTokens(
+    safeAddress: string,
+    tokenAddress: string,
+    recipient: string,
+    amount: string,
+    options?: ContractOptions
+  ): Promise<GnosisExecTx> {
+    let from = options?.from ?? (await this.layer2Web3.eth.getAccounts())[0];
+    let token = new this.layer2Web3.eth.Contract(ERC20ABI as AbiItem[], tokenAddress);
+    let safeBalance = new BN(await token.methods.balanceOf(safeAddress).call());
+    if (safeBalance.lt(new BN(amount))) {
+      throw new Error(
+        `Safe does not have enough balance to transfer tokens. The token ${tokenAddress} balance of safe ${safeAddress} is ${fromWei(
+          safeAddress.toString()
+        )}, amount to transfer ${fromWei(amount)}`
+      );
+    }
+    let payload = await this.transferTokenPayload(tokenAddress, recipient, amount);
+    let estimate = await gasEstimate(this.layer2Web3, safeAddress, tokenAddress, '0', payload, 0, tokenAddress);
+    if (estimate.lastUsedNonce == null) {
+      estimate.lastUsedNonce = -1;
+    }
+    let signatures = await sign(
+      this.layer2Web3,
+      tokenAddress,
+      0,
+      payload,
+      0,
+      estimate.safeTxGas,
+      estimate.dataGas,
+      estimate.gasPrice,
+      estimate.gasToken,
+      ZERO_ADDRESS,
+      toBN(estimate.lastUsedNonce + 1),
+      from,
+      safeAddress
+    );
+    let result = await executeTransaction(
+      this.layer2Web3,
+      safeAddress,
+      tokenAddress,
+      0,
+      payload,
+      0,
+      estimate.safeTxGas,
+      estimate.dataGas,
+      estimate.gasPrice,
+      toBN(estimate.lastUsedNonce + 1).toString(),
+      signatures,
+      estimate.gasToken,
+      ZERO_ADDRESS
+    );
+    return result;
+  }
+
+  private async transferTokenPayload(tokenAddress: string, recipient: string, amount: string): Promise<string> {
+    let token = new this.layer2Web3.eth.Contract(ERC20ABI as AbiItem[], tokenAddress);
+
+    return token.methods.transfer(recipient, amount).encodeABI();
   }
 }
